chore: Move logic to view model
